Allow to specify a different user & group than cmsuser in the config
This required moving a couple of things around as mkdir, in util.py,
needed to know that config value but util.py was an import of conf.py
and thus couldn't itself import conf.py in turn.
