+++ cms/io/service.py
@@ -273,7 +273,7 @@ class Service(object):
         backdoor_sock = _socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
         backdoor_sock.setblocking(0)
         backdoor_sock.bind(backdoor_path)
-        user = pwd.getpwnam("cmsuser")
+        user = pwd.getpwnam(config.cmsuser)
         # We would like to also set the user to "cmsuser" but only root
         # can do that. Therefore we limit ourselves to the group.
         os.chown(backdoor_path, os.getuid(), user.pw_gid)
