+++ cms/util.py
@@ -40,13 +40,10 @@ from collections import namedtuple
 
 import gevent.socket
 
+from cms import ServiceCoord, ConfigError, async_config, config
+
 
 logger = logging.getLogger(__name__)
-
-
-class ConfigError(Exception):
-    """Exception for critical configuration errors."""
-    pass
 
 
 def mkdir(path):
@@ -60,7 +57,7 @@ def mkdir(path):
         os.mkdir(path)
         try:
             os.chmod(path, 0o770)
-            cmsuser_gid = grp.getgrnam('cmsuser').gr_gid
+            cmsuser_gid = grp.getgrnam(config.cmsuser).gr_gid
             os.chown(path, -1, cmsuser_gid)
         except OSError as error:
             os.rmdir(path)
@@ -94,41 +91,6 @@ def utf8_decoder(value):
                 pass
 
     raise TypeError("Not a string.")
-
-
-class Address(namedtuple("Address", "ip port")):
-    def __repr__(self):
-        return "%s:%d" % (self.ip, self.port)
-
-
-class ServiceCoord(namedtuple("ServiceCoord", "name shard")):
-    """A compact representation for the name and the shard number of a
-    service (thus identifying it).
-
-    """
-    def __repr__(self):
-        return "%s,%d" % (self.name, self.shard)
-
-
-class Config(object):
-    """This class will contain the configuration for the
-    services. This needs to be populated at the initilization stage.
-
-    The *_services variables are dictionaries indexed by ServiceCoord
-    with values of type Address.
-
-    Core services are the ones that are supposed to run whenever the
-    system is up.
-
-    Other services are not supposed to run when the system is up, or
-    anyway not constantly.
-
-    """
-    core_services = {}
-    other_services = {}
-
-
-async_config = Config()
 
 
 def get_safe_shard(service, provided_shard):
